// beauty-core-backend/src/modules/automacoes/eventos/tipo-evento-sistema.enum.ts
export enum TipoEventoSistema {
  AGENDAMENTO_CRIADO = 'AGENDAMENTO_CRIADO',
  AGENDAMENTO_CONFIRMADO = 'AGENDAMENTO_CONFIRMADO',
  AGENDAMENTO_CANCELADO = 'AGENDAMENTO_CANCELADO',
  AGENDAMENTO_CONCLUIDO = 'AGENDAMENTO_CONCLUIDO',

  PONTOS_ADICIONADOS = 'PONTOS_ADICIONADOS',
  PONTOS_RESGATADOS = 'PONTOS_RESGATADOS',
  BENEFICIO_LIBERADO = 'BENEFICIO_LIBERADO',
  NIVEL_ALTERADO = 'NIVEL_ALTERADO',

  PACOTE_CRIADO = 'PACOTE_CRIADO',
  PACOTE_FINALIZADO = 'PACOTE_FINALIZADO',
  PACOTE_VENCIDO = 'PACOTE_VENCIDO',

  MOVIMENTACAO_FINANCEIRA = 'MOVIMENTACAO_FINANCEIRA',

  COMISSAO_GERADA = 'COMISSAO_GERADA',
  COMISSAO_PAGA = 'COMISSAO_PAGA',

  CLIENTE_ANIVERSARIANTE = 'CLIENTE_ANIVERSARIANTE',
  CLIENTE_CADASTRADO = 'CLIENTE_CADASTRADO',

  SERVICO_CADASTRADO = 'SERVICO_CADASTRADO',
}